Add files via upload
In this project me (Andrea Edson Lorenzoni) and (Pedro Ferreira) we worked together to create all the scripts required by the assignment Appendix A. We tested them on postgresql and we create the SuperheroDB.

There are three main tables you need to create, this is Superhero, Assistant, and Power.
Superhero has: Autoincremented integer Id, Name, Alias, Origin. 
Assistant has: Autoincremented integer Id, Name.
Power has: Autoincremented integer Id, Name, Description.
REQUIREMENT: Create a script called 01_tableCreate.sql that contains statements to create each of these tables and 
setup their primary keys. 
NOTE: At this point in time, no relationships are to be configured aside from primary keys. Datatypes should be up to you, 
and what you think is best.
3) Table relationships
Here we are going to include some relationships between the tables. The ALTER SQL keyword will be used to change the 
existing tables to add the keys. A description of the relationships can be seen below:
• One Superhero can have multiple assistants, one assistant has one superhero they assist.
• One Superhero can have many powers, and one power can be present on many Superheroes.
Firstly, to setup the Superhero-assistant relationship, we need to add a foreign key. This represents the superhero this 
assistant is linked to (SuperheroId in Assistant), and ALTER any tables needed to cater for this foreign key, and to setup 
the constraint.
REQUIREMENT: Create a script called 02_relationshipSuperheroAssistant.sql that contains statements to ALTER any 
tables needed to add the foreign key and setup the constraint to configure the described relationship between 
Superhero and assistant. 
Finally, to setup the Superhero-power relationship, we need to add a linking table. The name of this table is up to you 
but should be based on convention. This table is there purely for linking, meaning it needs to contain only two fields, 
which are both foreign keys and a composite primary key.
REQUIREMENT: Create a script called 03_relationshipSuperheroPower.sql that contains statements to create the linking 
table. This script should contain any ALTER statements needed to set up the foreign key constraints between the linking 
tables and the Superhero and Power tables.
4) Inserting data
Now it is time to populate the database using INSERT statements. Firstly, we should add superheroes.
REQUIREMENT: Create a script called 04_insertSuperheroes.sql that inserts three new superheroes into the database.
Then we need to add assistant data.
REQUIREMENT: Create a script called 05_insertAssistants.sql that inserts three assistants, decide on which superheroes 
these can assist.
Finally, we need to add powers. This requires you to add the powers first, then associate them with Superheroes.
REQUIREMENT: Create a script called 06_powers.sql that inserts four powers. Then the script needs to give the 
superheroes powers. Try have one superhero with multiple powers and one power that is used by multiple superheroes, 
to demonstrate the many-to-many.
5) Updating data
Pick one superhero to update their data.
REQUIREMENT: Create a script called 07_updateSuperhero.sql where you can update a superheroes name. Pick any 
superhero to do this with.
6) Deleting data
Referential integrity will stop us from deleting any records which have got relationships used. Meaning, a superhero 
cannot be deleted if it is used in the assistant table.
REQUIREMENT: Create a script called 08_deleteAssistant.sql where you can delete any assistant. You can delete that 
assistant by name (his name must be unique to avoid deleing multiple assistants), this is done to ease working with 
autoincremented numbers – my PC may skip a number or two.
NOTE: The scripts should be able to be run in order with no errors to create a database with the requirements stated. That 
is how I will initially check for functionality, then I will investigate each statement for specifics.